Delete some muhash fuzz inputs
The current muhash run takes 4060 seconds https://cirrus-ci.com/task/5011520649166848?command=ci#L4779
This is more than 1 hour spent on a target that is unlikely to be modified and thus break.

It seems fine to let the fuzz farms in the background tackle this target and avoid CI timeouts by
removing some fuzz inputs.